move create_react_agent_prompt to agent.py to make register.py callable without langchain.  update imports.
