+++ src/aiq/agent/react_agent/register.py
@@ -15,7 +15,6 @@
 
 import logging
 
-from langchain_core.prompts import ChatPromptTemplate
 from pydantic import Field
 
 from aiq.agent.base import AGENT_LOG_PREFIX
@@ -60,49 +59,15 @@ class ReActAgentWorkflowConfig(FunctionBaseConfig, name="react_agent"):
         default=None, description="Additional instructions to provide to the agent in addition to the base prompt.")
 
 
-def create_react_agent_prompt(config: ReActAgentWorkflowConfig) -> ChatPromptTemplate:
-    """
-    Create a ReAct Agent prompt from the config.
-
-    Args:
-        config (ReActAgentWorkflowConfig): The config to use for the prompt.
-
-    Returns:
-        ChatPromptTemplate: The ReAct Agent prompt.
-    """
-    # the ReAct Agent prompt comes from prompt.py, and can be customized there or via config option system_prompt.
-    from langchain_core.prompts import MessagesPlaceholder
-
-    from aiq.agent.react_agent.prompt import SYSTEM_PROMPT
-    from aiq.agent.react_agent.prompt import USER_PROMPT
-
-    from .agent import ReActAgentGraph
-
-    if config.system_prompt:
-        _prompt_str = config.system_prompt
-    else:
-        _prompt_str = SYSTEM_PROMPT
-
-    if config.additional_instructions:
-        _prompt_str += f" {config.additional_instructions}"
-
-    valid_prompt = ReActAgentGraph.validate_system_prompt(_prompt_str)
-    if not valid_prompt:
-        logger.exception("%s Invalid system_prompt", AGENT_LOG_PREFIX)
-        raise ValueError("Invalid system_prompt")
-    prompt = ChatPromptTemplate([("system", _prompt_str), ("user", USER_PROMPT),
-                                 MessagesPlaceholder(variable_name='agent_scratchpad', optional=True)])
-    return prompt
-
-
 @register_function(config_type=ReActAgentWorkflowConfig, framework_wrappers=[LLMFrameworkEnum.LANGCHAIN])
 async def react_agent_workflow(config: ReActAgentWorkflowConfig, builder: Builder):
     from langchain.schema import BaseMessage
     from langchain_core.messages import trim_messages
     from langgraph.graph.graph import CompiledGraph
 
-    from .agent import ReActAgentGraph
-    from .agent import ReActGraphState
+    from aiq.agent.react_agent.agent import ReActAgentGraph
+    from aiq.agent.react_agent.agent import ReActGraphState
+    from aiq.agent.react_agent.agent import create_react_agent_prompt
 
     prompt = create_react_agent_prompt(config)
